test/ca/sheridancollege/project/MainTest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit4TestClass.java to edit this template
 */
package ca.sheridancollege.project;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author jiveshmalhotra
 */
public class MainTest {
    
    public MainTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of main method, of class Main.
     */

    /**
     * Test of isInteger method, of class Main.
     */
    @Test
    public void testIsIntegerGood() {
        System.out.println("isInteger");
        String input = "2";
        boolean expResult = true;
        boolean result = Main.isInteger(input);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }
    @Test
    public void testIsIntegerBad() {
        System.out.println("isNotInteger");
        String input = "gcwzghfedwsjqhv";
        boolean expResult = false;
        boolean result = Main.isInteger(input);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }
    @Test
    public void testIsIntegerBoundary() {
        System.out.println("isNotInteger");
        String input = "";
        boolean expResult = false;
        boolean result = Main.isInteger(input);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }
    
}
